+++ frontend/src/components/cfp-form/index.tsx
@@ -38,6 +38,9 @@ import {
   PublicProfileCard,
 } from "../public-profile-card";
 import { TagsSelect } from "../tags-select";
+import { AboutYouSection } from "./about-you-section";
+import { AvailabilitySection } from "./availability-section";
+import { ProposalSection } from "./proposal-section";
 
 export type CfpFormFields = ParticipantFormFields & {
   type: string;
@@ -53,6 +56,7 @@ export type CfpFormFields = ParticipantFormFields & {
   previousTalkVideo: string;
   shortSocialSummary: string;
   acceptedPrivacyPolicy: boolean;
+  selectedAvailabilities: { [time: number]: null | boolean };
 };
 
 export type SubmissionStructure = {
@@ -81,29 +85,6 @@ type Props = {
   error: ApolloError | undefined;
   data: SendSubmissionMutation | UpdateSubmissionMutation;
 };
-
-const SPEAKER_LEVEL_OPTIONS = [
-  {
-    value: "",
-    disabled: true,
-    messageId: "cfp.selectSpeakerLevel",
-  },
-  {
-    disabled: false,
-    value: "new",
-    messageId: "cfp.speakerLevel.new",
-  },
-  {
-    disabled: false,
-    value: "intermediate",
-    messageId: "cfp.speakerLevel.intermediate",
-  },
-  {
-    disabled: false,
-    value: "experienced",
-    messageId: "cfp.speakerLevel.experienced",
-  },
-];
 
 const filterOutInactiveLanguages = (
   value: MultiLingualInputType,
@@ -149,7 +130,7 @@ export const CfpForm = ({
     },
   );
 
-  const { textarea, radio, select, checkbox, url, raw } = formOptions;
+  const { checkbox } = formOptions;
 
   const { data: conferenceData } = useCfpFormQuery({
     variables: {
@@ -292,8 +273,6 @@ export const CfpForm = ({
     }
   }, []);
 
-  const inputPlaceholder = useTranslatedMessage("input.placeholder");
-
   const hasValidationErrors =
     submissionData?.mutationOp.__typename === "SendSubmissionErrors";
 
@@ -326,274 +305,37 @@ export const CfpForm = ({
       submissionData!.mutationOp.errors[key]) ||
     [];
 
+  const onChangeAvailability = (date, choice) => {
+    formState.setField("selectedAvailabilities", {
+      ...formState.values.selectedAvailabilities,
+      [date.getTime()]: choice,
+    });
+  };
+
   return (
     <form onSubmit={submitSubmission} autoComplete="off">
-      <MultiplePartsCard>
-        <CardPart contentAlign="left">
-          <Heading size={3}>
-            <FormattedMessage id="cfp.youridea" />
-          </Heading>
-        </CardPart>
-        <CardPart background="milk" contentAlign="left">
-          <Grid cols={1} gap="medium">
-            <InputWrapper
-              required={true}
-              title={<FormattedMessage id="cfp.choosetype" />}
-              description={<FormattedMessage id="cfp.choosetypeDescription" />}
-            >
-              <VerticalStack gap="small">
-                {conferenceData!.conference.submissionTypes.map((type) => (
-                  <label key={type.id}>
-                    <HorizontalStack gap="small" alignItems="center">
-                      <Checkbox
-                        {...radio("type", type.id)}
-                        required={true}
-                        size="small"
-                      />
-                      <Text weight="strong" size={2}>
-                        {type.name}
-                      </Text>
-                    </HorizontalStack>
-                  </label>
-                ))}
-              </VerticalStack>
-            </InputWrapper>
-
-            <InputWrapper
-              required={true}
-              title={<FormattedMessage id="cfp.languagesLabel" />}
-              description={<FormattedMessage id="cfp.languagesDescription" />}
-            >
-              <HorizontalStack gap="small">
-                {conferenceData!.conference.languages.map((language) => (
-                  <label key={language.code}>
-                    <HorizontalStack gap="small" alignItems="center">
-                      <Checkbox
-                        size="small"
-                        {...checkbox("languages", language.code)}
-                      />
-                      <Text weight="strong" size={2}>
-                        {language.name}
-                      </Text>
-                    </HorizontalStack>
-                  </label>
-                ))}
-              </HorizontalStack>
-            </InputWrapper>
-
-            <InputWrapper
-              required={true}
-              title={<FormattedMessage id="cfp.title" />}
-              description={<FormattedMessage id="cfp.titleDescription" />}
-            >
-              <MultiLingualInput
-                {...raw("title")}
-                languages={formState.values.languages}
-              >
-                <Input
-                  required={true}
-                  maxLength={100}
-                  errors={getErrors("validationTitle")}
-                  placeholder={inputPlaceholder}
-                />
-              </MultiLingualInput>
-            </InputWrapper>
-
-            <InputWrapper
-              required={true}
-              title={<FormattedMessage id="cfp.elevatorPitchLabel" />}
-              description={
-                <FormattedMessage id="cfp.elevatorPitchDescription" />
-              }
-            >
-              <MultiLingualInput
-                {...raw("elevatorPitch")}
-                languages={formState.values.languages}
-              >
-                <Textarea
-                  required={true}
-                  maxLength={300}
-                  rows={6}
-                  errors={getErrors("validationElevatorPitch")}
-                  placeholder={inputPlaceholder}
-                />
-              </MultiLingualInput>
-            </InputWrapper>
-
-            <InputWrapper
-              required={true}
-              title={<FormattedMessage id="cfp.lengthLabel" />}
-              description={<FormattedMessage id="cfp.lengthDescription" />}
-            >
-              <Select
-                {...select("length")}
-                required={true}
-                errors={getErrors("validationDuration")}
-              >
-                <FormattedMessage id="cfp.selectDuration">
-                  {(txt) => (
-                    <option value="" disabled={true}>
-                      {txt}
-                    </option>
-                  )}
-                </FormattedMessage>
-                {allowedDurations!.map((d) => (
-                  <option key={d.id} value={d.id}>
-                    {d.name}
-                  </option>
-                ))}
-              </Select>
-            </InputWrapper>
-
-            <InputWrapper
-              required={true}
-              title={<FormattedMessage id="cfp.audienceLevelLabel" />}
-              description={
-                <FormattedMessage id="cfp.audienceLevelDescription" />
-              }
-            >
-              <Select
-                {...select("audienceLevel")}
-                required={true}
-                errors={getErrors("validationAudienceLevel")}
-              >
-                <FormattedMessage id="cfp.selectAudience">
-                  {(txt) => (
-                    <option value="" disabled={true}>
-                      {txt}
-                    </option>
-                  )}
-                </FormattedMessage>
-                {conferenceData!.conference.audienceLevels.map((a) => (
-                  <option key={a.id} value={a.id}>
-                    {a.name}
-                  </option>
-                ))}
-              </Select>
-            </InputWrapper>
-
-            <InputWrapper
-              required={true}
-              title={<FormattedMessage id="cfp.tagsLabel" />}
-              description={<FormattedMessage id="cfp.tagsDescription" />}
-            >
-              <TagsSelect
-                tags={formState.values.tags || []}
-                onChange={(tags: { value: string }[]) => {
-                  formState.setField(
-                    "tags",
-                    tags.map((t) => t.value),
-                  );
-                }}
-              />
-            </InputWrapper>
-
-            <InputWrapper
-              required={true}
-              title={<FormattedMessage id="cfp.abstractLabel" />}
-              description={<FormattedMessage id="cfp.abstractDescription" />}
-            >
-              <MultiLingualInput
-                {...raw("abstract")}
-                languages={formState.values.languages}
-              >
-                <Textarea
-                  required={true}
-                  maxLength={5000}
-                  rows={6}
-                  errors={getErrors("validationAbstract")}
-                  placeholder={inputPlaceholder}
-                />
-              </MultiLingualInput>
-            </InputWrapper>
-
-            <InputWrapper
-              title={<FormattedMessage id="cfp.notesLabel" />}
-              description={<FormattedMessage id="cfp.notesDescription" />}
-            >
-              <Textarea
-                {...textarea("notes")}
-                maxLength={1000}
-                rows={4}
-                errors={getErrors("validationNotes")}
-                placeholder={inputPlaceholder}
-              />
-            </InputWrapper>
-
-            <InputWrapper
-              required={false}
-              title={<FormattedMessage id="cfp.shortSocialSummaryLabel" />}
-              description={
-                <FormattedMessage id="cfp.shortSocialSummaryDescription" />
-              }
-            >
-              <Textarea
-                {...textarea("shortSocialSummary")}
-                required={false}
-                maxLength={128}
-                rows={2}
-                errors={getErrors("validationShortSocialSummary")}
-                placeholder={inputPlaceholder}
-              />
-            </InputWrapper>
-          </Grid>
-        </CardPart>
-      </MultiplePartsCard>
+      <ProposalSection
+        allowedDurations={allowedDurations}
+        formState={formState}
+        getErrors={getErrors}
+        conferenceData={conferenceData}
+        formOptions={formOptions}
+      />
 
       <Spacer size="medium" />
 
-      <MultiplePartsCard>
-        <CardPart contentAlign="left">
-          <Heading size={3}>
-            <FormattedMessage id="cfp.aboutYou" />
-          </Heading>
-        </CardPart>
-        <CardPart background="milk" contentAlign="left">
-          <Grid cols={1} gap="medium">
-            <Text size={2}>
-              <FormattedMessage id="cfp.aboutYouDescription" />
-            </Text>
-            <InputWrapper
-              required={true}
-              title={<FormattedMessage id="cfp.speakerLevel" />}
-              description={
-                <FormattedMessage id="cfp.speakerLevelDescription" />
-              }
-            >
-              <Select
-                {...select("speakerLevel")}
-                required={true}
-                errors={getErrors("validationSpeakerLevel")}
-              >
-                {SPEAKER_LEVEL_OPTIONS.map(({ value, disabled, messageId }) => (
-                  <FormattedMessage id={messageId} key={messageId}>
-                    {(copy) => (
-                      <option disabled={disabled} value={value}>
-                        {copy}
-                      </option>
-                    )}
-                  </FormattedMessage>
-                ))}
-              </Select>
-            </InputWrapper>
+      <AboutYouSection getErrors={getErrors} formOptions={formOptions} />
 
-            <InputWrapper
-              title={<FormattedMessage id="cfp.previousTalkVideo" />}
-              description={
-                <FormattedMessage id="cfp.previousTalkVideoDescription" />
-              }
-            >
-              <Input
-                {...url("previousTalkVideo")}
-                required={false}
-                maxLength={2048}
-                errors={getErrors("validationPreviousTalkVideo")}
-                placeholder={inputPlaceholder}
-              />
-            </InputWrapper>
-          </Grid>
-        </CardPart>
-      </MultiplePartsCard>
+      <Spacer size="medium" />
+
+      <AvailabilitySection
+        onChangeAvailability={onChangeAvailability}
+        selectedAvailabilities={formState.values.selectedAvailabilities}
+        selectedDuration={allowedDurations.find(
+          (duration) => duration.id === formState.values.length,
+        )}
+        conferenceData={conferenceData}
+      />
 
       <Spacer size="medium" />
 
@@ -612,7 +354,6 @@ export const CfpForm = ({
 
       {!submission && (
         <>
-          {" "}
           <label>
             <HorizontalStack gap="small" alignItems="center">
               <Checkbox
